+++ tests/e2e-tests/src/specs/regression/expenses/crud-credit-card-expenses.spec.ts
@@ -66,7 +66,9 @@ test.afterAll(async () => {
 	} catch {
 		console.log('No user to be deleted');
 	} finally {
-		await graphqlService.deleteCategory(categoryId);
+		if(graphqlService) {
+			await graphqlService.deleteCategory(categoryId);
+		}
 		await page.close();
 	}
 });
